+++ cmd/pint/lint.go
@@ -75,11 +75,14 @@ func actionLint(c *cli.Context) error {
 	}
 
 	slog.Info("Finding all rules to check", slog.Any("paths", paths))
-	finder := discovery.NewGlobFinder(paths, git.NewPathFilter(
-		config.MustCompileRegexes(meta.cfg.Parser.Include...),
-		config.MustCompileRegexes(meta.cfg.Parser.Exclude...),
-		config.MustCompileRegexes(meta.cfg.Parser.Relaxed...),
-	))
+	finder := discovery.NewGlobFinder(
+		paths,
+		git.NewPathFilter(
+			config.MustCompileRegexes(meta.cfg.Parser.Include...),
+			config.MustCompileRegexes(meta.cfg.Parser.Exclude...),
+			config.MustCompileRegexes(meta.cfg.Parser.Relaxed...),
+		),
+		parseSchema(meta.cfg.Parser.Schema))
 	entries, err := finder.Find()
 	if err != nil {
 		return err