+++ internal/discovery/discovery_test.go
@@ -24,7 +24,7 @@ func (r failingReader) Read(_ []byte) (int, error) {
 
 func TestReadRules(t *testing.T) {
 	mustParse := func(offset int, s string) parser.Rule {
-		p := parser.NewParser(false)
+		p := parser.NewParser(false, parser.PrometheusSchema)
 		r, err := p.Parse([]byte(strings.Repeat("\n", offset) + s))
 		if err != nil {
 			panic(fmt.Sprintf("failed to parse rule:\n---\n%s\n---\nerror: %s", s, err))
@@ -338,7 +338,7 @@ groups:
 			fmt.Sprintf("rPath=%s sPath=%s strict=%v title=%s", tc.reportedPath, tc.sourcePath, tc.isStrict, tc.title),
 			func(t *testing.T) {
 				r := tc.sourceFunc(t)
-				entries, err := readRules(tc.reportedPath, tc.sourcePath, r, tc.isStrict)
+				entries, err := readRules(tc.reportedPath, tc.sourcePath, r, tc.isStrict, parser.PrometheusSchema)
 				if tc.err != "" {
 					require.EqualError(t, err, tc.err)
 				} else {