+++ internal/discovery/glob.go
@@ -9,18 +9,21 @@ import (
 	"path/filepath"
 
 	"github.com/cloudflare/pint/internal/git"
+	"github.com/cloudflare/pint/internal/parser"
 )
 
-func NewGlobFinder(patterns []string, filter git.PathFilter) GlobFinder {
+func NewGlobFinder(patterns []string, filter git.PathFilter, schema parser.Schema) GlobFinder {
 	return GlobFinder{
 		patterns: patterns,
 		filter:   filter,
+		schema:   schema,
 	}
 }
 
 type GlobFinder struct {
 	patterns []string
 	filter   git.PathFilter
+	schema   parser.Schema
 }
 
 func (f GlobFinder) Find() (entries []Entry, err error) {
@@ -66,7 +69,7 @@ func (f GlobFinder) Find() (entries []Entry, err error) {
 		if err != nil {
 			return nil, err
 		}
-		el, err := readRules(fp.target, fp.path, fd, !f.filter.IsRelaxed(fp.target))
+		el, err := readRules(fp.target, fp.path, fd, !f.filter.IsRelaxed(fp.target), f.schema)
 		if err != nil {
 			fd.Close()
 			return nil, fmt.Errorf("invalid file syntax: %w", err)